Initial introduction of the POJOAsserter
Introduced a class to wrap the logic of asserting a POJO's getters and
setters and extract it out of AssertUtils. This is done as an
intermediate phase to allow configuring the behavior of these
assertions.

Note that the current commit does not allow this yet, but only sets up
the initial refactoring needed to do this.