Create loggers by classes, not class names
Commit 01afed10e9f347036d518089c83c357c28c5fa8a ported BPUnit from using
java.util.logging directly to the more portable and pluggable SLF4J.
That commit, however, did not change the style in which the Logger was
created, and left the invocation via a String, as was done beforehand.
This commit finishes the job and initializes the Logger from a Class
object, which is more elegant, and potentially allows catching stupid
mistakes early on if the "slf4j.detectLoggerNameMismatch" system property is
set to "true".