Reimplement SeedableRandom#nextString(int, boolean)
SeedableRandom#nexString(int, boolean)'s implementation was borrowed
from the oVirt Engine's org.ovirt.engine.core.utils.RandomUtils.
This implementation would randomize a byte[] of the given length and
then create a String from it. However, examining the documentation
for String(byte[], Charset) [1] states that:
"""
The length of the new String is a function of the charset, and hence
may not be equal to the length of the byte array.
"""

This means that the original implementation did not provide the
expected behavior of producing a random String of the given length.
While the test suite passes despite the faulty implementation,
TravisCI's run with Java 7 (both Oracle's and OpenJDK's JDKs) actually
fails due to it [2].

This patch re-implements this method in a straight-forward way by
randomizing the chars that make up the String instead of randomizing
bytes.

[1] https://docs.oracle.com/javase/7/docs/api/java/lang/String.html#String(byte[],%20java.nio.charset.Charset)
[2] https://travis-ci.org/mureinik/bpunit/builds/143535815