LoggingBehavior: Add ability to suppress exception
The default logging behavior will print a stacktrace of the exception
passed to it, which is often not that useful. This patch adds the
ability to ignore the exception and not log its stacktrace.

This is done by adding a boolean parameter to the LoggingBehavior's
constructor. The no-arg constructor is maintained and passes true to
the overloaded constructor, for backwards compatibility.